App/DiagramScreen => Refactored Code | v1.7
