+++ App/Indoor-React-Native/api/server.js
@@ -5,9 +5,8 @@ const app = express()
 const port = 3000
 
 app.get('/', (req, res) => {
-  res.send({
-    altura: 'alto',
-    movilidad: 'movible'
+  res.json({
+    name: 'puto'
   })
 })
